+++ packages/payload-plugin-scheduling/CHANGELOG.md
@@ -1,3 +1,10 @@
+# @focus-reactive/payload-plugin-scheduling [1.2.0](https://github.com/focusreactive/payload-plugins/compare/@focus-reactive/payload-plugin-scheduling@1.1.0...@focus-reactive/payload-plugin-scheduling@1.2.0) (2026-04-30)
+
+
+### Features
+
+* **payload-plugin-scheduling:** add time granularity config ([ca55edf](https://github.com/focusreactive/payload-plugins/commit/ca55edff2116f933682e12e5535e8e6090d88f0a))
+
 # @focus-reactive/payload-plugin-scheduling [1.1.0](https://github.com/focusreactive/payload-plugins/compare/@focus-reactive/payload-plugin-scheduling@1.0.1...@focus-reactive/payload-plugin-scheduling@1.1.0) (2026-03-31)